==> 02_PY_statistic_test.ipynb <==
# Preparing folder where data located. If file located in local drive
import os
print(os.getcwd())

#Load related libraries to do analysis
import pandas as pd #To work with dataset
import numpy as np #Math library
import seaborn as sns #Graph library that use matplot in background
import matplotlib.pyplot as plt #to plot some parameters in seaborn
%matplotlib inline

## III.2.1 Persiapan data dan impor ke Python

# library panda untuk menggunakan data frame
import pandas as pd

# Pengaturan folder tempat data / Google Colab

# Impor data dari excel xls ke Python
df = pd.read_excel('input/german_credit_easy.xls')

# correlation semua
print(df.corr())
# correlation untuk 2 faktor saja
print(np.corrcoef(df['installment_as_income_perc'], df['default'])) 
# Menghitung korelasi antara variabel Age dengan default
print(df['default'].corr(df['age']))

# Uji coba hipotesis apakah ada perbedaan nilai rata-rata variabel Age dengan default (gagal bayar dan tidak gagal bayar)
from scipy import stats
stats.ttest_ind( df[df.default == 0].age,
                 df[df.default == 1].age, nan_policy='omit')
