==> 03_PY_data_preparation.ipynb <==
#Load related libraries to do analysis
import pandas as pd #To work with dataset
import numpy as np #Math library

## III.2.1 Persiapan data dan impor ke Python

# library panda untuk menggunakan data frame
import pandas as pd

# Pengaturan folder tempat data / Google Colab

# Impor data dari excel xls ke Python
df = pd.read_excel('input/german_credit_easy.xls')

## III.3.1 Penanganan untuk data yang hilang (missing values)

# Membuat kopi dari df dan menyimpan sebagai df_prepare
df_prepare = df.copy()

# Koding berikut menggantikan missing value dari semua variabel dengan
# nilai yang paling sering muncul (modus)

for column in df_prepare.columns:
        df_prepare[column].fillna(df_prepare[column].mode()[0], inplace=True)

## III.3.3 Pembuatan variabel yang baru (Feature Engineering) 

df_prepare['cicilan_per_bulan'] = df_prepare['credit_amount'] / df_prepare['duration_in_month']

df_prepare.info()

# Create new variable using def function
# Note that this will be slow with big dataframe
def category_duration(row):
    if (row['duration_in_month'] < 0):
        val = '0.Negative?'
    elif (row['duration_in_month'] >= 24) & (row['duration_in_month'] <= 36):
        val = '1.24<x<=36mth'
    elif (row['duration_in_month'] < 24) | (row['duration_in_month'] > 36):
        val = '2.<24or>36mth'
    else:
        val = '3.WrongNumber' #default value
    return val
df_prepare['category_duration'] = df_prepare.apply(category_duration, axis = 1)

# Create new variable using numpy operation
# This will be very fast even with big dataframe
conditions =    [
                (df_prepare['duration_in_month'] < 0),
                ((df_prepare['duration_in_month'] >= 24) & (df_prepare['duration_in_month'] <= 36)),
                ((df_prepare['duration_in_month'] < 24) | (df_prepare['duration_in_month'] > 36))
                ]
choices =       [
                '0.Negative?',
                '1.24<x<=36mth',
                '2.<24or>36mth'
                ]
df_prepare['category_duration2'] = np.select(conditions,choices, default=np.nan) #nan is default value

## III.3.4 Transformasi variabel kategori menjadi numerik (One-Hot Encoding) 

# Koding untuk one hot encoding untuk semua variabel kategori

# Langkah pertama adalah membuang variabel ID karena ini adalah identifikasi nasabah
df_prepare.drop(['ID'], axis = 1, inplace = True)

# Memisahkan variabel kategori dengan variabel numerik
dataset_dummies = df_prepare.select_dtypes(exclude = ['float64','int64','int'])
dataset_int = df_prepare.select_dtypes(include=['float64','int64','int'])

# Memisahkan target/dependent variabel dari dataset dengan variabel kategori
target = df_prepare['default']
dataset_int.drop(['default'], axis = 1, inplace = True)

# Menggunakan function get_dummies dari panda untuk melakukan one hot encoding
dataset_dummies_df = pd.get_dummies(dataset_dummies,drop_first = True)
    
# Menggabungkan dataset variabel kategori dari one hot encoding dengan variabel numerik
dataset = pd.concat([dataset_dummies_df, dataset_int], axis = 1)
feature_name = dataset.columns
df_encoded = pd.concat([dataset, target], axis = 1)

df_encoded.to_excel('input/german_credit_encoded.xlsx', index=False)

# Subset dataframe by criteria
example1 = df_encoded[df_encoded.default == 0]

# Only select some columns from dataframe
example1 = df_prepare[['default','category_duration2']]

df_encoded.to_pickle('output/df_encoded.pkl')
