==> 08_PY_modeling_automl.ipynb <==
# Auto ML in Python with German Credit data

Automated Machine Learning, often abbreviated as AutoML, is an emerging field in which the process of building machine learning models to model data is automated. AutoML has the capability to make modelling easier and more accessible for everyone.
If you’re interested in checking out AutoML, these four Python libraries are the way to go. 

Based on data of German Credit that already clean up and prepared

### Resources and References
https://towardsdatascience.com/4-python-automl-libraries-every-data-scientist-should-know-680ff5d6ad08

#Load related libraries to do analysis
import pandas as pd #To work with dataset
import numpy as np #Math library
import seaborn as sns #Graph library that use matplot in background
import matplotlib.pyplot as plt #to plot some parameters in seaborn
%matplotlib inline
from sklearn.model_selection import train_test_split

## Machine Learning Part

# machine learning library
from sklearn.linear_model import LogisticRegression
from sklearn.svm import SVC, LinearSVC
import statsmodels.api as sm

#to check performance of model
from sklearn.metrics import accuracy_score
from sklearn.metrics import confusion_matrix
from sklearn.metrics import classification_report

#plot
import matplotlib.pyplot as plt

# libraries for automl
#autosklearn only works in linux
#import autosklearn as ask

# Impor data dari excel xls german credit versi encoded (sudah dalam angka semua) ke Python
final_data = pd.read_excel('input/german_credit_encoded.xlsx')

x = final_data.drop(['default'],axis=1)
y = final_data['default']

x_train, x_test, y_train, y_test = train_test_split(x,y,test_size=0.3, random_state=101)

#autosklearn only works in linux
import autosklearn as ask
#ask.regression.AutoSklearnRegressor() for regression tasks
model = ask.classification.AutoSklearnClassifier(ensemble_size=10, #size of the end ensemble (minimum is 1)
                                                 time_left_for_this_task=120, #the number of seconds the process runs for
                                                 per_run_time_limit=30) #maximum seconds allocated per model
model.fit(x_train, y_train) #begin fitting the search model
print(model.sprint_statistics()) #print statistics for the search
y_predictions = model.predict(x_test) #get predictions from the model

import tpot
pipeline_optimizer = tpot.TPOTClassifier(generations=5, #number of iterations to run the training
                                         population_size=20, #number of individuals to train
                                         cv=5) #number of folds in StratifiedKFold
pipeline_optimizer.fit(x_train, y_train) #fit the pipeline optimizer - can take a long time
print(pipeline_optimizer.score(x_test, y_test)) #print scoring for the pipeline
pipeline_optimizer.export('output/tpot_exported_pipeline.py') #export the pipeline - in Python code!
