==> 20_PY_fannie_mae_prepare.ipynb <==
# Assume df_acquisition is the original acquisition file

# Assume df_performance is the original performance fiel

# This coding will get a sample of 10,000 only customer taken by random

#Load related libraries to do analysis
import pandas as pd #To work with dataset
import numpy as np #Math library

df_acquisition = pd.read_table('input/Acquisition_2000Q1.txt', sep = '|', header = None
                        ,names = ["LOAN_ID", "ORIG_CHN", "Seller.Name", "ORIG_RT", "ORIG_AMT", "ORIG_TRM", "ORIG_DTE","FRST_DTE", "OLTV",                                       "OCLTV", "NUM_BO", "DTI", "CSCORE_B", "FTHB_FLG", "PURPOSE", "PROP_TYP","NUM_UNIT", "OCC_STAT", "STATE",                                      "ZIP_3", "MI_PCT", "Product.Type", "CSCORE_C", "MI_TYPE", "RELOCATION_FLG"])
df_performance = pd.read_table('input/Performance_2000Q1.txt', sep = '|', header = None
                        ,names = ["LOAN_ID", "Monthly.Rpt.Prd", "Servicer.Name", "LAST_RT", "LAST_UPB", "Loan.Age", "Months.To.Legal.Mat"
                                  ,"Adj.Month.To.Mat", "Maturity.Date", "MSA", "Delq.Status", "MOD_FLAG", "Zero.Bal.Code","ZB_DTE", "LPI_DTE"                                   ,"FCC_DTE","DISP_DT", "FCC_COST", "PP_COST", "AR_COST", "IE_COST", "TAX_COST", "NS_PROCS","CE_PROCS"                                          ,"RMW_PROCS", "O_PROCS", "NON_INT_UPB", "PRIN_FORG_UPB_FHFA", "REPCH_FLAG", "PRIN_FORG_UPB_OTH"
                                  ,"TRANSFER_FLAG"])

# Randomly choose 10,000 loan_id
df_acquisition_sample = df_acquisition.sample(n = 1500) 
df_acquisition_sample.info()

# Create final performance table
df_performance_sample = pd.merge(df_acquisition_sample, df_performance, how='inner', left_on='LOAN_ID', right_on='LOAN_ID')

df_performance_sample.info()
df_performance_sample.to_csv('output/fm2000Q1_performance_sample.txt',sep='|', index=False)
df_acquisition_sample.info()
df_acquisition_sample.to_csv('output/fm2000Q1_acquisition_sample.txt',sep='|', index=False)
