==> 40_PY_Scorecard_Docs.ipynb <==
# Application Scorecard: Model Development and Documentation Template
This is an example of documentation explaining the model development. Often required as part of review

## [Name of Bank/Multifinance/Fintect] [Product of Secured/Unsecured] [Application or Behaviour] Scorecard
Scorecard version 0.1
Document Version Control: Scorecard_Draft_Indonesia_Unsecured_A_Score_2020_v0.1
| Date | Version |
| ---- | ------- |
| Dec 15, 2020 | 0.1 |
